refactor(hypr): split utility modules and components
